% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_noisy_frame}
\alias{read_noisy_frame}
\title{Load a container saved by \code{noisyvalue.io.save()} as a tibble}
\usage{
read_noisy_frame(path)
}
\arguments{
\item{path}{Path to the JSON file written by \code{noisyvalue.io.save()}. Must
have been a \code{pandas.DataFrame} when saved.}
}
\description{
Load a container saved by \code{noisyvalue.io.save()} as a tibble
}
